amadeus_test_mcp/requirements.txt
google-adk>=1.20.0
google-cloud-aiplatform[adk,agent_engines]
mcp>=1.0.0
amadeus>=12.0.0
pydantic>=2.11.0
httpx>=0.28.0
python-dotenv>=1.0.0
